Update platform support for the role
Since this role is no longer reliant on Python 2 we can enable all
platforms that would be supported by default in one of our Ansible
roles. However, Amazon Linux 2023 is disabled due to a dependency
conflict between the pre-installed gnupg2-minimal package and this role
attempting to install the full gnupg2 package.